Started rearchitecting
